Upgrade CI: Ubuntu, MongoDB Server, Extension, Symfony (#2757)
Since mongodb extension v2.0.0 has been released, we need to explicitly install the version 1.21.0 to test this version
Ubuntu 20.04 Image is disabled by Github, upgraded to 22.04 for unit tests that need to setup MongoDB Server. 24.04 for other tests.
MongoDB 5.0 is not supported by on Ubuntu 22.04, so the minimum MongoDB server tested is 6.0.
Adding test on MongoDB 8.0
Adding test on Symfony 6.4: lowest dependencies installs 5.4 and highest installs 7.2+